requirements.txt
matplotlib
pandas
pyyaml
jupyter
qgrid
pyperclip
selenium
ipywidgets
sphinx
numpydoc
sphinxcontrib-napoleon
